Construct transactions directly on wire format
This is using fixed cost models and maximum execution units, but
constructs our transaction using the game redeemer and datum directly in
the transaction's cbor format.

This brings down execution of `hydraSend` from 30ms+ to 1-3ms!

Idea and implementation credit goes to @Quantumplation

I just added the script hash and a few bytes.